requirements.txt
streamlit
streamlit-extras
groq
python-dotenv
supabase
fastapi
uvicorn
pydantic
requests
